requirements.txt
aniso8601==7.0.0
asgiref==3.8.1
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.4.0
django-filter==24.2
django-graphql-auth==0.3.16
django-graphql-jwt==0.4.0
django-jazzmin==3.0.0
django_csp==3.8
environ==1.0
graphene==2.1.9
graphene-django==2.16.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.7
pillow==10.4.0
promise==2.3
PyJWT==2.8.0
reportlab==4.2.2
requests==2.32.3
Rx==1.6.3
singledispatch==4.1.0
six==1.16.0
sqlparse==0.5.0
stripe==10.2.0
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
